codegen, symbol_mangling: prefix/move LLVM CFI functionality to avoid confusion.
