+++ compiler/rustc_codegen_gcc/src/builder.rs
@@ -783,11 +783,11 @@ impl<'a, 'gcc, 'tcx> BuilderMethods<'a, 'tcx> for Builder<'a, 'gcc, 'tcx> {
         // TODO(antoyo)
     }
 
-    fn type_metadata(&mut self, _function: RValue<'gcc>, _typeid: String) {
+    fn llvm_cfi_type_metadata(&mut self, _function: RValue<'gcc>, _typeid: String) {
         // Unsupported.
     }
 
-    fn typeid_metadata(&mut self, _typeid: String) -> RValue<'gcc> {
+    fn llvm_cfi_typeid_metadata(&mut self, _typeid: String) -> RValue<'gcc> {
         // Unsupported.
         self.context.new_rvalue_from_int(self.int_type, 0)
     }
